Move LoginPage to web/pages/login subfolder
