+++ gsa/CMakeLists.txt
@@ -489,7 +489,7 @@ set (GSA_JS_SRC_FILES
      ${GSA_SRC_DIR}/src/web/pages/hosts/dashboard/topologydisplay.js
      ${GSA_SRC_DIR}/src/web/pages/hosts/dashboard/vulnscoredisplay.js
      ${GSA_SRC_DIR}/src/web/pages/ldap/ldappage.js
-     ${GSA_SRC_DIR}/src/web/pages/loginpage.js
+     ${GSA_SRC_DIR}/src/web/pages/login/loginpage.js
      ${GSA_SRC_DIR}/src/web/pages/notes/component.js
      ${GSA_SRC_DIR}/src/web/pages/notes/details.js
      ${GSA_SRC_DIR}/src/web/pages/notes/detailspage.js
